+++ integrations/github/github/webhook/webhook_processors/dependabot_webhook_processor.py
@@ -6,7 +6,7 @@ from github.helpers.utils import (
     enrich_with_repository,
     enrich_with_organization,
 )
-from github.clients.client_factory import create_github_client
+from github.clients.client_factory import create_github_client_for_org
 from port_ocean.core.handlers.port_app_config.models import ResourceConfig
 from port_ocean.core.handlers.webhook.webhook_event import (
     EventPayload,
@@ -70,7 +70,7 @@ class DependabotAlertWebhookProcessor(BaseRepositoryWebhookProcessor):
                 updated_raw_results=[], deleted_raw_results=[alert]
             )
 
-        rest_client = create_github_client()
+        rest_client = await create_github_client_for_org(organization)
         exporter = RestDependabotAlertExporter(rest_client)
 
         data_to_upsert = await exporter.get_resource(